183 crear interfaz del post (#193)
Comentado el post

+++ src/AppForSEII2526.API/Controllers/HerramientasController.cs
@@ -62,7 +62,7 @@ namespace AppForSEII2526.API.Controllers
         [HttpGet("DisponiblesReparacion")]
         [ProducesResponseType(typeof(List<HerramientaRepaDTO>), StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<ActionResult<List<HerramientaRepaDTO>>> GetHerramientasDisponiblesParaReparar(string? NombreHerramienta , int? DiaReparacion)
+        public async Task<ActionResult<List<HerramientaRepaDTO>>> GetHerramientasDisponiblesParaReparar(string? NombreHerramienta , float? DiaReparacion)
         {
             if (_context.Herramienta == null)
             {
